pyproject.toml
[tool.poetry]
name = "eagle-eye-discord-bot"
version = "0.1.0"
description = ""
authors = ["willis81808"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.4.0"
openai = "^1.52.2"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
asyncio = "^3.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
